pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-tenants-router"
version = "1.0.0"
description = "Multi-tenant database routing for Django with Redis caching"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = ["django", "multi-tenant", "database", "router", "redis", "saas"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP",
]

dependencies = [
    "Django>=4.0",
    "cryptography==46.0.7"
]

[project.optional-dependencies]
redis = ["redis>=4.0"]
drf = ["djangorestframework>=3.13"]
jwt = ["djangorestframework-simplejwt>=5.0"]
dev = [
    "pytest>=7.0",
    "pytest-django>=4.5",
    "pytest-cov>=4.0",
    "black",
    "isort",
    "flake8",
    "mypy",
    "django-stubs",
]
all = [
    "redis>=4.0",
    "djangorestframework>=3.13",
    "djangorestframework-simplejwt>=5.0",
]

[project.urls]
Homepage = "https://github.com/Sushil2308/django-tenants-router"
Documentation = "https://django-tenants-router.readthedocs.io"
Repository = "https://github.com/Sushil2308/django-tenants-router"
"Bug Tracker" = "https://github.com/Sushil2308/django-tenants-router/issues"

[tool.setuptools.packages.find]
include = ["django_tenants_router*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["test_*.py"]
python_classes = ["Test*", "*Test", "*Tests"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["django_tenants_router"]
omit = ["*/migrations/*", "*/tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
known_django = ["django"]
known_first_party = ["django_tenants_router"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
